Worked on Terrain
Added a terrain map and did a basic mock up of the path and the crater with trees marking the path. Not sure how big we want the map to be so feel free to shrink if necessary. Same with the crater. Also added asset packs to use for props